IFS-182 - WIP fixes for persistor services

+++ ifs-data-layer/ifs-data-service/src/main/java/org/innovateuk/ifs/competition/transactional/template/GuidanceRowTemplatePersistorService.java
@@ -38,6 +38,8 @@ public class GuidanceRowTemplatePersistorService implements BaseChainedTemplateP
 
     public void cleanForPrecedingEntity(FormInput formInput) {
         List<GuidanceRow> scoreRows = formInput.getGuidanceRows();
-        guidanceRowRepository.delete(scoreRows);
+        if(scoreRows.size() > 0) {
+            guidanceRowRepository.delete(scoreRows);
+        }
     }
 }
